feat(like-dislike): add like dislike feature

+++ internal/service/memberships/login.go
@@ -26,7 +26,7 @@ func (s *service) Login(request memberships.LoginRequest) (string, error) {
 		return "", errors.New("email and password not match")
 	}
 
-	accessToken, err := jwt.CreateToken(int64(userDetail.ID), userDetail.Username, s.cfg.Service.SecretKey)
+	accessToken, err := jwt.CreateToken(userDetail.ID, userDetail.Username, s.cfg.Service.SecretKey)
 	if err != nil {
 		log.Error().Err(err).Msg("failed to create JWT Token")
 		return "", err
